% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esa.R
\name{ESA.list}
\alias{ESA.list}
\title{List all available SDMX tables}
\usage{
ESA.list(update = FALSE)
}
\arguments{
\item{update}{Boolean. If true, then the table will be recreated from internet. Otherwise it's loaded from memory}
}
\description{
Get the list of all tables available in Statistical Office of Estonia
}
\examples{
\dontrun{
ESA.list() # Lists all the tables
ESA.list(update = TRUE)  # Updates the list
}
}
