% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esa.R
\name{ESA.metadata}
\alias{ESA.metadata}
\title{Get Metadata}
\usage{
ESA.metadata(table)
}
\arguments{
\item{table}{A string, representing the name of the table}
}
\value{
This function creates a dataframes with list of variables and dimensions in the table
}
\description{
Gets table structure information from Estonian Statistical Office
}
\examples{
ESA.metadata("IA02")
}
