% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esa.R
\name{ESA.monthly}
\alias{ESA.monthly}
\title{Creates monthly time series from the data
Use as}
\usage{
ESA.monthly(table, year = "obsTime", month = "DIM3", name = "obsDate")
}
\arguments{
\item{table}{dataframe of the raw data, returned from the statistical offic}

\item{year}{optional column name for the year data. Defaults to obsTime}

\item{month}{optional column name for the month data. Defaults to DIM3}

\item{name}{optional column name for the newly created date field. Defaults to obsDate}
}
\value{
dataframe with the newly created time column
}
\description{
Dpendencies:
}
\examples{
\dontrun{
ESA.monthly(data)
ESA.monthly(data, month = 'DIM2')
}
}
