% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esa.R
\name{ESA.query}
\alias{ESA.query}
\title{Dependencies}
\usage{
ESA.query(table, restrictions = NA, start = NA, end = NA)
}
\arguments{
\item{table}{A string of a table name. e.g IA02}

\item{restrictions}{A string or vector of an optional restrictions list.}

\item{start}{An optional year number for start of the data}

\item{end}{An optional year number for end of the data}
}
\value{
A string of url for database query
}
\description{
Requests data from Statistical Office of Estonia (ESA)
}
\examples{
\dontrun{
ESA.url("IA02")  # Returns a whole CPI table
ESA.url("IA02", '1')  # Returns a CPI table for id = 1 Total
ESA.url("IA02", c('1+3','1+2'), 1995, 2000) # Returns CPI table for Total and Alcohol, January and February in dates 1995-2000
}
}
