% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esa.R
\name{ESA.titles}
\alias{ESA.titles}
\title{Table title}
\usage{
ESA.titles(table)
}
\arguments{
\item{table}{A string of a table name. e.g IA02}
}
\value{
dataframe of official titles of the table from Statistics Estonia
}
\description{
Returns title of the table
}
\examples{
ESA.titles("IA02")
}
